--- a/src/app/(dashboard)/followers/page.tsx
+++ b/src/app/(dashboard)/followers/page.tsx
@@ -1,4 +1,10 @@
+import { generateSiteMetadata } from '$lib/configs/site';
 import { FollowersView } from '$lib/features/followers/views/followers';
+
+export const metadata = generateSiteMetadata({
+	title: 'Followers',
+	robots: { follow: false, index: false }
+});
 
 export default function FollowersPage() {
 	return <FollowersView />;