--- a/src/app/(dashboard)/whitelist/page.tsx
+++ b/src/app/(dashboard)/whitelist/page.tsx
@@ -1,4 +1,10 @@
+import { generateSiteMetadata } from '$lib/configs/site';
 import { WhitelistView } from '$lib/features/whitelist/views/whitelist';
+
+export const metadata = generateSiteMetadata({
+	title: 'Whitelist',
+	robots: { follow: false, index: false }
+});
 
 export default function WhitelistPage() {
 	return <WhitelistView />;