--- a/src/app/legal/terms/page.tsx
+++ b/src/app/legal/terms/page.tsx
@@ -1,0 +1,8 @@
+import { generateSiteMetadata } from '$lib/configs/site';
+import { TermsView } from '$lib/features/legal/views/terms';
+
+export const metadata = generateSiteMetadata({ title: 'Terms of Use' });
+
+export default function PrivacyPolicyPage() {
+	return <TermsView />;
+}