--- a/src/lib/components/ui/compose.tsx
+++ b/src/lib/components/ui/compose.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 export function Compose({
 	children,
 	components