--- a/src/lib/components/ui/skeleton.tsx
+++ b/src/lib/components/ui/skeleton.tsx
@@ -1,0 +1,13 @@
+import { cn } from '$lib/utils/ui';
+
+function Skeleton({ className, ...props }: React.ComponentProps<'div'>) {
+	return (
+		<div
+			data-slot='skeleton'
+			className={cn('bg-accent animate-pulse rounded-md', className)}
+			{...props}
+		/>
+	);
+}
+
+export { Skeleton };