--- a/src/lib/configs/repo/index.ts
+++ b/src/lib/configs/repo/index.ts
@@ -1,0 +1,1 @@
+export * from './repo.config';