--- a/src/lib/configs/repo/repo.config.ts
+++ b/src/lib/configs/repo/repo.config.ts
@@ -1,0 +1,20 @@
+export type TRepoMetadata = {
+	owner: {
+		email: string;
+		name: string;
+		username: string;
+	};
+	repo: string;
+};
+
+/**
+ * Metadata information about the repository owner and the repository itself.
+ */
+export const repoMetadata: TRepoMetadata = {
+	owner: {
+		email: 'me@gustavomorinaga.dev',
+		name: 'Gustavo Morinaga',
+		username: 'gustavomorinaga'
+	},
+	repo: 'github-unfollow-checker'
+};