--- a/src/lib/features/auth/components/account-details/index.ts
+++ b/src/lib/features/auth/components/account-details/index.ts
@@ -1,0 +1,1 @@
+export * from './account-details.component';