--- a/src/lib/features/auth/components/account-dropdown-menu/index.ts
+++ b/src/lib/features/auth/components/account-dropdown-menu/index.ts
@@ -1,0 +1,1 @@
+export * from './account-dropdown-menu.component';