--- a/src/lib/features/auth/components/account/index.ts
+++ b/src/lib/features/auth/components/account/index.ts
@@ -1,1 +1,0 @@
-export * from './account.component';