- [x] Add [@rescale/slim](https://www.rescale.build/docs/packages/slim) as compression suite for JavaScript objects
- [x] Optimize API calls compressing response data
- [x] Optimize Local Storage compressing the saved data

## API Benchmark Results

| Metric         | Before          | After            | Improvement              |
|----------------|-----------------|------------------|--------------------------|
| Size           | 589 kB          | 63.9 kB          | 89.15% smaller           |
| Response Time  | 874 ms          | 729 ms           | 16.6% faster             |